+++ astro/rollup.config.mjs
@@ -2,8 +2,7 @@ import resolve from "@rollup/plugin-node-resolve";
 import commonjs from "@rollup/plugin-commonjs";
 import json from "@rollup/plugin-json";
 import replace from "@rollup/plugin-replace";
-// import terser from "@rollup/plugin-terser";
-import sourcemaps from "rollup-plugin-sourcemaps";
+import terser from "@rollup/plugin-terser";
 import dts from "rollup-plugin-dts";
 
 import { createRequire } from "node:module";
@@ -28,13 +27,10 @@ let plugins = [
   json(),
 ];
 
-// Check if this is a dev or a prod build
-if (NODE_ENV === "development") {
-  // Add sourcemaps for dev builds
-  plugins.push(sourcemaps);
-} else {
+// Check if this is a prod build
+if (NODE_ENV !== "development") {
   // Add terser for prod builds
-  // plugins.push(terser());
+  plugins.push(terser());
 }
 
 export default [
@@ -42,9 +38,9 @@ export default [
     // This is to rollup the shell lib
     input: "dist/main.js",
     output: {
-      sourcemap: NODE_ENV === "development",
       file: "dist/astro.mjs",
       format: "es",
+      sourcemap: true,
     },
     external: ["@bs-core/shell", "astro/app"],
     plugins,