+++ plugins/ldap/rollup.config.mjs
@@ -3,7 +3,6 @@ import commonjs from "@rollup/plugin-commonjs";
 import json from "@rollup/plugin-json";
 import replace from "@rollup/plugin-replace";
 import terser from "@rollup/plugin-terser";
-import sourcemaps from "rollup-plugin-sourcemaps";
 import dts from "rollup-plugin-dts";
 
 import { createRequire } from "node:module";
@@ -28,10 +27,8 @@ let plugins = [
   json(),
 ];
 
-if (NODE_ENV === "development") {
-  // Add sourcemaps for dev builds
-  plugins.push(sourcemaps);
-} else {
+// Check if this is a prod build
+if (NODE_ENV !== "development") {
   // Add terser for prod builds
   plugins.push(terser());
 }
@@ -40,9 +37,9 @@ export default [
   {
     input: "dist/main.js",
     output: {
-      sourcemap: NODE_ENV === "development",
       file: "dist/plugin.mjs",
       format: "es",
+      sourcemap: true,
     },
     external: ["ldapjs", "@bs-core/shell"],